+++ src/__tests__/ghost.test.js
@@ -31,7 +31,7 @@ describe('Ghost Chat', () => {
   })
 
   beforeEach(async () => {
-    jest.setTimeout(20000)
+    jest.setTimeout(30000)
   })
 
   it('creates chat correctly', async () => {
@@ -64,23 +64,32 @@ describe('Ghost Chat', () => {
       ipfs2 = await utils.initIPFS(12)
     })
 
-    it('creates second chat correctly', async (done) => {
+    it('creates second chat correctly', async () => {
+      // checks if chat2 joined properly
+      const c1Promise = new Promise(resolve => {
+        chat.on('user-joined', async (_event, did, peerId) => {
+          expect(_event).toEqual('joined')
+          const members = await chat.listMembers()
+          expect(members).toEqual(expect.arrayContaining([DID2]))
+          resolve()
+        })
+      })
       chat2 = new GhostThread(CHAT_NAME, { ipfs: ipfs2 });
+      const c2Promise = new Promise(resolve => {
+        chat2.on('user-joined', async (_event, did, peerId) => {
+          expect(_event).toEqual('joined')
+          const members2 = await chat2.listMembers()
+          expect(members2).toEqual(expect.arrayContaining([DID1]))
+          resolve()
+        })
+      })
       chat2._set3id(THREEID2_MOCK)
       expect(chat2._name).toEqual(CHAT_NAME)
       expect(chat2._3id).toEqual(THREEID2_MOCK)
       expect(chat2.listMembers()).toBeDefined()
       expect(chat2.getPosts()).toBeDefined()
-
-      // checks if chat2 joined properly
-      chat.on('user-joined', async (_event, did, peerId) => {
-        expect(_event).toEqual('joined')
-        const members = await chat.listMembers()
-        const members2 = await chat2.listMembers()
-        expect(members).toEqual(expect.arrayContaining([DID2]))
-        expect(members2).toEqual(expect.arrayContaining([DID1]))
-        done()
-      })
+      await c1Promise
+      await c2Promise
     })
 
     it('chat2 should catch broadcasts from chat', async (done) => {
@@ -197,12 +206,12 @@ describe('Ghost Chat', () => {
 
     afterAll(async () => {
       await chat3.close()
-      await utils.stopIPFS(ipfs3, 12)
+      return utils.stopIPFS(ipfs3, 12)
     })
   })
 
   afterAll(async () => {
     await chat1.close()
-    await utils.stopIPFS(ipfs, 11)
+    return utils.stopIPFS(ipfs, 11)
   })
 })