+++ src/__tests__/idWallet.integration.js
@@ -81,9 +81,9 @@ describe('Integration Test: IdentityWallet', () => {
   })
 
   afterAll(async () => {
-    await pubsub.disconnect()
-    await testUtils.stopIPFS(ipfs1, 9)
-    await testUtils.stopIPFS(ipfs2, 10)
+    // await pubsub.disconnect()
+    // await testUtils.stopIPFS(ipfs1, 9)
+    // return testUtils.stopIPFS(ipfs2, 10)
   })
 
   it('should create and auth correctly when idWallet is passed', async () => {
@@ -98,7 +98,7 @@ describe('Integration Test: IdentityWallet', () => {
     pubAddr = box.public._db.address.toString()
     privAddr = box.private._db.address.toString()
     pubState = await box.public.all()
-    await box.close()
+    return box.close()
   })
 
   it('should get same state on second open and auth', async () => {
@@ -110,11 +110,22 @@ describe('Integration Test: IdentityWallet', () => {
     expect(await box.public.all()).toEqual(pubState)
 
     expect(box.replicator.rootstore._oplog.values.length).toEqual(3)
+    const writePromise = new Promise(resolve => {
+      let firstWrite = true
+      const writeHandler = () => {
+        if (!firstWrite) {
+          box.replicator.rootstore.events.off('write', writeHandler)
+          resolve()
+        }
+        firstWrite = false
+      }
+      box.replicator.rootstore.events.on('write', writeHandler)
+    })
     idWallet.addAuthMethod(AUTH_1)
-    await testUtils.delay(1200)
+    await writePromise
     expect(box.replicator.rootstore._oplog.values.length).toEqual(5)
 
-    await box.close()
+    return box.close()
   })
 
   it('should get same state on create and auth with IdentityWallet opened using first authSecret', async () => {
@@ -127,11 +138,22 @@ describe('Integration Test: IdentityWallet', () => {
     expect(await box.public.all()).toEqual(pubState)
 
     expect(box.replicator.rootstore._oplog.values.length).toEqual(5)
+    const writePromise = new Promise(resolve => {
+      let firstWrite = true
+      const writeHandler = () => {
+        if (!firstWrite) {
+          box.replicator.rootstore.events.off('write', writeHandler)
+          resolve()
+        }
+        firstWrite = false
+      }
+      box.replicator.rootstore.events.on('write', writeHandler)
+    })
     idWallet.addAuthMethod(AUTH_2)
-    await testUtils.delay(1200)
+    await writePromise
     expect(box.replicator.rootstore._oplog.values.length).toEqual(7)
 
-    await box.close()
+    return box.close()
   })
 
   it('should get same state on create and auth with IdentityWallet opened using second authSecret', async () => {
@@ -142,6 +164,6 @@ describe('Integration Test: IdentityWallet', () => {
     await box.auth([], opts)
     await box.syncDone
     expect(await box.public.all()).toEqual(pubState)
-    await box.close()
+    return box.close()
   })
 })