+++ src/__tests__/space.test.js
@@ -61,7 +61,7 @@ describe('Space', () => {
     //expect(rootstoreMock.add).toHaveBeenCalledWith({ type: 'space', DID: threeIdMock.getSubDID(NAME1), odbAddress:'/orbitdb/myodbaddr' })
     expect(threeIdMock.isAuthenticated).toHaveBeenCalledWith([NAME1])
     expect(threeIdMock.authenticate).toHaveBeenCalledWith([NAME1])
-    await syncDonePromise
+    return syncDonePromise
   })
 
   it('should return directly if space already opened', async () => {
@@ -252,7 +252,7 @@ describe('Space', () => {
       expect(t1._load).toHaveBeenCalledTimes(1)
       expect(t1._setIdentity).toHaveBeenCalledTimes(0)
       // function for autosubscribing works as intended
-      await Thread.mock.calls[0][4](threadAddress)
+      return Thread.mock.calls[0][4](threadAddress)
     })
   })
 })