+++ src/__tests__/thread.test.js
@@ -49,7 +49,7 @@ describe('Thread', () => {
 
   beforeAll(async () => {
     ipfs = await utils.initIPFS(4)
-    const identity = await Identities.createIdentity({ id: 'nullid' })
+    const identity = await Identities.createIdentity({ id: 'nullid', identityKeysPath: './tmp/odbIdentityKeys-thread' })
     orbitdb = await OrbitDB.createInstance(ipfs, {
       directory:'./tmp/orbitdb4',
       identity
@@ -459,12 +459,12 @@ describe('Thread', () => {
 
     afterAll(async () => {
       await orbitdb2.stop()
-      await utils.stopIPFS(ipfs2, 5)
+      return utils.stopIPFS(ipfs2, 5)
     })
   })
 
   afterAll(async () => {
     await orbitdb.stop()
-    await utils.stopIPFS(ipfs, 4)
+    return utils.stopIPFS(ipfs, 4)
   })
 })